split controls to seperate file
